+++ tests/test_support_pandas.py
@@ -1,7 +1,9 @@
 import re
 import sys
 
+import pandas as pd
 import pytest
+from pandas._testing import assert_equal
 from pueblo.testing.pandas import makeTimeDataFrame
 from sqlalchemy.exc import ProgrammingError
 
@@ -14,6 +16,24 @@ INSERT_RECORDS = 42
 # Create dataframe, to be used as input data.
 df = makeTimeDataFrame(nper=INSERT_RECORDS, freq="S")
 df["time"] = df.index
+
+float_double_data = {
+    "col_1": [19556.88, 629414.27, 51570.0, 2933.52, 20338.98],
+    "col_2": [
+        15379.920000000002,
+        1107140.42,
+        8081.999999999999,
+        1570.0300000000002,
+        29468.539999999997,
+    ],
+}
+float_double_df = pd.DataFrame.from_dict(float_double_data)
+
+float_nan_df = pd.DataFrame.from_dict(
+    {
+        "col_1": [float("nan"), float("inf"), float("-inf")],
+    }
+)
 
 
 @pytest.mark.skipif(
@@ -113,3 +133,58 @@ def test_table_kwargs_unknown(cratedb_service):
                 "passed to [ALTER | CREATE] TABLE statement]"
             )
         )
+
+
+@pytest.mark.skipif(
+    sys.version_info < (3, 8), reason="Feature not supported on Python 3.7 and earlier"
+)
+@pytest.mark.skipif(
+    SA_VERSION < SA_2_0, reason="Feature not supported on SQLAlchemy 1.4 and earlier"
+)
+def test_float_special_values(cratedb_service):
+    """
+    Validate CrateDB's handling of special float values: NaN, Inf, -Inf.
+
+    CrateDB stores all of them as NULL, so they all read back as NaN in pandas.
+    """
+    tablename = "pandas_float_special"
+    engine = cratedb_service.database.engine
+    float_nan_df.to_sql(
+        tablename,
+        engine,
+        if_exists="replace",
+        index=False,
+    )
+    cratedb_service.database.run_sql(f"REFRESH TABLE {tablename}")
+    df_load = pd.read_sql_table(tablename, engine)
+    assert df_load["col_1"].isna().all()
+
+
+@pytest.mark.skipif(
+    sys.version_info < (3, 8), reason="Feature not supported on Python 3.7 and earlier"
+)
+@pytest.mark.skipif(
+    SA_VERSION < SA_2_0, reason="Feature not supported on SQLAlchemy 1.4 and earlier"
+)
+def test_float_double(cratedb_service):
+    """
+    Validate I/O with floating point numbers, specifically DOUBLE types.
+
+    Motto: Do not lose precision when DOUBLE is required.
+    """
+    tablename = "pandas_double"
+    engine = cratedb_service.database.engine
+    float_double_df.to_sql(
+        tablename,
+        engine,
+        if_exists="replace",
+        index=False,
+    )
+    cratedb_service.database.run_sql(f"REFRESH TABLE {tablename}")
+    df_load = pd.read_sql_table(tablename, engine)
+
+    before = float_double_df.sort_values(by="col_1", ignore_index=True)
+    after = df_load.sort_values(by="col_1", ignore_index=True)
+
+    pd.options.display.float_format = "{:.12f}".format
+    assert_equal(before, after, check_exact=True)